use `usize` to store index and use functional programming style to simplify code (#23)

+++ crates/allay-compiler/src/ast.rs
@@ -162,7 +162,7 @@ pub enum TopLevel {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum GetField {
-    Index(i32),
+    Index(usize),
     Name(String),
 }
